Move Fetch rank implementation to common CF package
